Move error into own file
