+++ src/xcode/ENA/ENA/Source/Services/CoronaTestService/CoronaTestService.swift
@@ -6,28 +6,6 @@ import Foundation
 import OpenCombine
 import UIKit
 
-enum CoronaTestServiceError: LocalizedError, Equatable {
-	case responseFailure(URLSession.Response.Failure)
-	case unknownTestResult
-	case testExpired
-	case noRegistrationToken
-	case noCoronaTestOfRequestedType
-
-	var errorDescription: String? {
-		switch self {
-		case let .responseFailure(responseFailure):
-			return responseFailure.errorDescription
-		case .noRegistrationToken:
-			return AppStrings.ExposureSubmissionError.noRegistrationToken
-		case .testExpired:
-			return AppStrings.ExposureSubmission.qrCodeExpiredAlertText
-		default:
-			Log.error("\(self)", log: .api)
-			return AppStrings.ExposureSubmissionError.defaultError
-		}
-	}
-}
-
 // swiftlint:disable:next type_body_length
 class CoronaTestService {
